+++ page_v2.go
@@ -3,6 +3,7 @@ package goparquet
 import (
 	"bytes"
 	"context"
+	"hash/crc32"
 	"io"
 
 	"github.com/fraugster/parquet-go/parquet"
@@ -72,7 +73,7 @@ func (dp *dataPageReaderV2) init(dDecoder, rDecoder getLevelDecoder, values getV
 	return nil
 }
 
-func (dp *dataPageReaderV2) read(r io.Reader, ph *parquet.PageHeader, codec parquet.CompressionCodec) error {
+func (dp *dataPageReaderV2) read(r io.Reader, ph *parquet.PageHeader, codec parquet.CompressionCodec, validateCRC bool) error {
 	// 1- Uncompressed size is affected by the level lens.
 	// 2- In page V2 the rle size is in header, not in level stream
 	if ph.DataPageHeaderV2 == nil {
@@ -99,30 +100,26 @@ func (dp *dataPageReaderV2) read(r io.Reader, ph *parquet.PageHeader, codec parq
 		}
 	}
 
-	// Its safe to call this {r,d}Decoder later, since the stream they operate on are in memory
+	dataPageBlock, err := readPageBlock(r, codec, ph.GetCompressedPageSize(), ph.GetUncompressedPageSize(), validateCRC, ph.Crc)
+	if err != nil {
+		return err
+	}
+
 	levelsSize := ph.DataPageHeaderV2.RepetitionLevelsByteLength + ph.DataPageHeaderV2.DefinitionLevelsByteLength
-	// read both level size
-	if levelsSize > 0 {
-		data := make([]byte, levelsSize)
-		n, err := io.ReadFull(r, data)
-		if err != nil {
-			return errors.Wrapf(err, "need to read %d byte but there was only %d byte", levelsSize, n)
-		}
 
-		if ph.DataPageHeaderV2.RepetitionLevelsByteLength > 0 {
-			if err := dp.rDecoder.init(bytes.NewReader(data[:int(ph.DataPageHeaderV2.RepetitionLevelsByteLength)])); err != nil {
-				return errors.Wrapf(err, "read repetition level failed")
-			}
-		}
-
-		if ph.DataPageHeaderV2.DefinitionLevelsByteLength > 0 {
-			if err := dp.dDecoder.init(bytes.NewReader(data[int(ph.DataPageHeaderV2.RepetitionLevelsByteLength):])); err != nil {
-				return errors.Wrapf(err, "read definition level failed")
-			}
+	if ph.DataPageHeaderV2.RepetitionLevelsByteLength > 0 {
+		if err = dp.rDecoder.init(bytes.NewReader(dataPageBlock[:int(ph.DataPageHeaderV2.RepetitionLevelsByteLength)])); err != nil {
+			return errors.Wrapf(err, "read repetition level failed")
 		}
 	}
 
-	reader, err := createDataReader(r, codec, ph.GetCompressedPageSize()-levelsSize, ph.GetUncompressedPageSize()-levelsSize)
+	if ph.DataPageHeaderV2.DefinitionLevelsByteLength > 0 {
+		if err = dp.dDecoder.init(bytes.NewReader(dataPageBlock[int(ph.DataPageHeaderV2.RepetitionLevelsByteLength):levelsSize])); err != nil {
+			return errors.Wrapf(err, "read definition level failed")
+		}
+	}
+
+	reader, err := newBlockReader(dataPageBlock[levelsSize:], codec, ph.GetCompressedPageSize()-levelsSize, ph.GetUncompressedPageSize()-levelsSize)
 	if err != nil {
 		return err
 	}
@@ -131,13 +128,14 @@ func (dp *dataPageReaderV2) read(r io.Reader, ph *parquet.PageHeader, codec parq
 }
 
 type dataPageWriterV2 struct {
-	col    *Column
-	schema SchemaWriter
-
-	codec      parquet.CompressionCodec
-	dictionary bool
 	dictValues []interface{}
+	col        *Column
+	schema     SchemaWriter
+	codec      parquet.CompressionCodec
 	page       *dataPage
+
+	dictionary bool
+	enableCRC  bool
 }
 
 func (dp *dataPageWriterV2) init(schema SchemaWriter, col *Column, codec parquet.CompressionCodec) error {
@@ -147,7 +145,7 @@ func (dp *dataPageWriterV2) init(schema SchemaWriter, col *Column, codec parquet
 	return nil
 }
 
-func (dp *dataPageWriterV2) getHeader(comp, unComp, defSize, repSize int, isCompressed bool, pageStats *parquet.Statistics, numRows int32) *parquet.PageHeader {
+func (dp *dataPageWriterV2) getHeader(comp, unComp, defSize, repSize int, isCompressed bool, pageStats *parquet.Statistics, numRows int32, crc32Checksum *int32) *parquet.PageHeader {
 	enc := dp.col.data.encoding()
 	if dp.dictionary {
 		enc = parquet.Encoding_RLE_DICTIONARY
@@ -156,7 +154,7 @@ func (dp *dataPageWriterV2) getHeader(comp, unComp, defSize, repSize int, isComp
 		Type:                 parquet.PageType_DATA_PAGE_V2,
 		UncompressedPageSize: int32(unComp + defSize + repSize),
 		CompressedPageSize:   int32(comp + defSize + repSize),
-		Crc:                  nil,
+		Crc:                  crc32Checksum,
 		DataPageHeaderV2: &parquet.DataPageHeaderV2{
 			NumValues:                  int32(dp.page.numValues) + int32(dp.page.nullValues),
 			NumNulls:                   int32(dp.page.nullValues),
@@ -210,9 +208,16 @@ func (dp *dataPageWriterV2) write(ctx context.Context, w io.Writer) (int, int, e
 	if err != nil {
 		return 0, 0, errors.Wrapf(err, "compressing data failed with %s method", dp.codec)
 	}
+
+	var crc32Checksum *int32
+	if dp.enableCRC {
+		v := int32(crc32.ChecksumIEEE(append(append(rep.Bytes(), def.Bytes()...), comp...)))
+		crc32Checksum = &v
+	}
+
 	compSize, unCompSize := len(comp), len(dataBuf.Bytes())
 	defLen, repLen := def.Len(), rep.Len()
-	header := dp.getHeader(compSize, unCompSize, defLen, repLen, dp.codec != parquet.CompressionCodec_UNCOMPRESSED, dp.page.stats, int32(dp.page.numRows))
+	header := dp.getHeader(compSize, unCompSize, defLen, repLen, dp.codec != parquet.CompressionCodec_UNCOMPRESSED, dp.page.stats, int32(dp.page.numRows), crc32Checksum)
 	if err := writeThrift(ctx, header, w); err != nil {
 		return 0, 0, err
 	}
@@ -228,10 +233,11 @@ func (dp *dataPageWriterV2) write(ctx context.Context, w io.Writer) (int, int, e
 	return compSize + defLen + repLen, unCompSize + defLen + repLen, writeFull(w, comp)
 }
 
-func newDataPageV2Writer(useDict bool, dictValues []interface{}, page *dataPage) pageWriter {
+func newDataPageV2Writer(useDict bool, dictValues []interface{}, page *dataPage, enableCRC bool) pageWriter {
 	return &dataPageWriterV2{
 		dictionary: useDict,
 		dictValues: dictValues,
 		page:       page,
+		enableCRC:  enableCRC,
 	}
 }